+++ database/migrations/2023_07_11_200310_add_liquid_license_domain_key_to_settings_two.php
@@ -21,8 +21,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('settings_two', function (Blueprint $table) {
-            $table->dropColumn('liquid_license_domain_key');
-        });
+        if (Schema::hasTable('settings_two')) {
+            Schema::table('settings_two', function (Blueprint $table) {
+                $table->dropColumn('liquid_license_domain_key');
+            });
+        }
     }
 };
