+++ database/migrations/2023_07_18_222043_add_image_storage_field_to_settings_two_table.php
@@ -20,8 +20,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('settings_two', function (Blueprint $table) {
-            $table->dropColumn('ai_image_storage');
-        });
+        if (Schema::hasTable('settings_two')) {
+            Schema::table('settings_two', function (Blueprint $table) {
+                $table->dropColumn('ai_image_storage');
+            });
+        }
     }
 };
