+++ database/migrations/2023_07_21_121324_options_to_settingstwo_table.php
@@ -21,8 +21,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('settings_two', function (Blueprint $table) {
-            $table->dropColumn('stablediffusion_default_language');
-        });
+        if (Schema::hasTable('settings_two')) {
+            Schema::table('settings_two', function (Blueprint $table) {
+                $table->dropColumn('stablediffusion_default_language');
+            });
+        }
     }
 };
