+++ database/migrations/2024_01_17_075234_add_image_url_to_extensions.php
@@ -22,7 +22,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('extensions', function (Blueprint $table) {
-            $table->dropColumn('image_url');
+            if (Schema::hasColumn('extensions', 'image_url')) {
+                $table->dropColumn('image_url');
+            }
         });
     }
 };
