+++ database/migrations/2024_01_18_001456_add_detail_to_extensions.php
@@ -22,7 +22,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('extensions', function (Blueprint $table) {
-            $table->dropColumn('detail');
+            if (Schema::hasColumn('extensions', 'detail')) {
+                $table->dropColumn('detail');
+            }
         });
     }
 };
