+++ database/migrations/2024_01_18_001457_add_licensed_to_extensions.php
@@ -22,7 +22,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('extensions', function (Blueprint $table) {
-            $table->dropColumn('licensed');
+            if (Schema::hasColumn('extensions', 'licensed')) {
+                $table->dropColumn('licensed');
+            }
         });
     }
 };
