+++ database/migrations/2024_03_08_112315_add_face_price_to_extensions_table.php
@@ -17,7 +17,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('extensions', function (Blueprint $table) {
-             $table->dropColumn('fake_price');
+            if (Schema::hasColumn('extensions', 'fake_price')) {
+                $table->dropColumn('fake_price');
+            }
         });
     }
 };
